pyproject.toml
[tool.poetry]
name = "drf-habits-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "drf_habits_tracker"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
celery = "^5.3.4"
flake8 = "^6.1.0"
drf-yasg = "^1.21.7"
django-redis = "^5.4.0"
django-cors-headers = "^4.3.0"
python-dotenv = "^1.0.0"
django-celery-beat = "^2.5.0"
djangorestframework-simplejwt = "^5.3.0"
requests = "^2.31.0"
coverage = "^7.3.2"
eventlet = "^0.33.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
